splitter ut metode for å angre send til beslutter fra gammel StegService
